+++ blueprint/billing-base/registrations.ts
@@ -7,6 +7,7 @@ import {
 } from '@forklaunch/blueprint-core';
 import { Metrics, metrics } from '@forklaunch/blueprint-monitoring';
 import { OpenTelemetryCollector } from '@forklaunch/core/http';
+import { FieldEncryptor } from '@forklaunch/core/persistence';
 import {
   ComplianceDataService,
   createConfigInjector,
@@ -140,6 +141,11 @@ const environmentConfig = configInjector.chain({
     lifetime: Lifetime.Singleton,
     type: string,
     value: getEnvVar('IAM_URL')
+  },
+  ENCRYPTION_KEY: {
+    lifetime: Lifetime.Singleton,
+    type: string,
+    value: getEnvVar('ENCRYPTION_KEY')
   }
 });
 
@@ -163,7 +169,12 @@ const runtimeDependencies = environmentConfig.chain({
   TtlCache: {
     lifetime: Lifetime.Singleton,
     type: RedisTtlCache,
-    factory: ({ REDIS_URL, OpenTelemetryCollector, OTEL_LEVEL }) =>
+    factory: ({
+      REDIS_URL,
+      OpenTelemetryCollector,
+      OTEL_LEVEL,
+      ENCRYPTION_KEY
+    }) =>
       new RedisTtlCache(
         60 * 60 * 1000,
         OpenTelemetryCollector,
@@ -173,6 +184,9 @@ const runtimeDependencies = environmentConfig.chain({
         {
           enabled: true,
           level: OTEL_LEVEL || 'info'
+        },
+        {
+          encryptor: new FieldEncryptor(ENCRYPTION_KEY)
         }
       )
   },